refactor(grn): align domain model and implement DTOs (Phase 19a)
- Make id required in GoodsReceivedNote and GrnItem interfaces
- Apply standard parsePositiveInt validation in GRN parsers
- Implement CreateGrnRequest and CreateGrnItemRequest DTOs
- Add JSON serialization helper createGrnToJson
- Refactor GRN types into dedicated domain files and export via index
