+++ types/grn/grn-item-create.ts
@@ -1,0 +1,6 @@
+export interface CreateGrnItemRequest {
+  materialId: number;
+  orderedQuantity: number;
+  receivedQuantity: number;
+  unitCost?: number;
+}
